+++ anathema-templates/src/token.rs
@@ -120,7 +120,8 @@ pub(crate) enum Kind {
     Value(Value),
     Op(Operator),
 
-    Decl,
+    Local,
+    Global,
 
     Eof,
 }
@@ -150,7 +151,8 @@ impl Display for Kind {
             Self::Indent(s) => write!(f, "<indent {s}>"),
             Self::Value(v) => write!(f, "<value {v}>"),
             Self::Op(o) => write!(f, "<op {o}>"),
-            Self::Decl => write!(f, "let"),
+            Self::Local => write!(f, "local"),
+            Self::Global => write!(f, "global"),
             Self::Eof => write!(f, "<Eof>"),
         }
     }